refactor: extracted module imports processing in
`ModuleTranslationState::new` and moved Miden ABI transformation
synthetic function definition and intrinsics to their modules.
