+++ frontend-wasm2/src/miden_abi/mod.rs
@@ -2,10 +2,24 @@ pub(crate) mod stdlib;
 pub(crate) mod transform;
 pub(crate) mod tx_kernel;
 
-use midenc_hir::{interner::Symbol, FunctionIdent, FunctionType, FxHashMap, Ident};
+use std::{cell::RefCell, rc::Rc};
+
+use midenc_dialect_hir::InstBuilder;
+use midenc_hir::{
+    dialects::builtin::{Function, ModuleBuilder, WorldBuilder},
+    interner::Symbol,
+    AbiParam, FunctionIdent, FunctionType, FxHashMap, Ident, Op, Signature, ValueRef,
+};
+use transform::transform_miden_abi_call;
 use tx_kernel::note;
 
-use crate::intrinsics;
+use crate::{
+    intrinsics,
+    module::{
+        function_builder_ext::{FunctionBuilderContext, FunctionBuilderExt, SSABuilderListener},
+        module_translation_state::CallableFunction,
+    },
+};
 
 pub(crate) type FunctionTypeMap = FxHashMap<&'static str, FunctionType>;
 pub(crate) type ModuleFunctionTypeMap = FxHashMap<&'static str, FunctionTypeMap>;
@@ -112,4 +126,76 @@ pub fn recover_imported_masm_module(wasm_module_id: String) -> Symbol {
         return wasm_module_id.into();
     };
     module_id.into()
+}
+
+/// Define a synthetic wrapper functon transforming parameters, calling the Miden ABI function
+/// (think written in MASM) and transforming result
+pub fn define_func_for_miden_abi_transformation(
+    world_builder: &mut WorldBuilder,
+    module_builder: &mut ModuleBuilder,
+    synth_func_id: FunctionIdent,
+    synth_func_sig: Signature,
+    import_func_id: FunctionIdent,
+) -> CallableFunction {
+    let import_ft = miden_abi_function_type(
+        import_func_id.module.as_symbol(),
+        import_func_id.function.as_symbol(),
+    );
+    let import_sig = Signature::new(
+        import_ft.params.into_iter().map(AbiParam::new),
+        import_ft.results.into_iter().map(AbiParam::new),
+    );
+    let mut func_ref = module_builder
+        .define_function(synth_func_id.function, synth_func_sig.clone())
+        .expect("failed to create an import function");
+    let mut func = func_ref.borrow_mut();
+    let span = func.name().span;
+    let context = func.as_operation().context_rc();
+    let func = func.as_mut().downcast_mut::<Function>().unwrap();
+    let func_ctx = Rc::new(RefCell::new(FunctionBuilderContext::new(context.clone())));
+    let mut op_builder =
+        midenc_hir2::OpBuilder::new(context).with_listener(SSABuilderListener::new(func_ctx));
+    let mut func_builder = FunctionBuilderExt::new(func, &mut op_builder);
+    let entry_block = func_builder.current_block();
+    func_builder.seal_block(entry_block); // Declare all predecessors known.
+    let args: Vec<ValueRef> = entry_block
+        .borrow()
+        .arguments()
+        .iter()
+        .copied()
+        .map(|ba| ba as ValueRef)
+        .collect();
+
+    let import_module_ref = if let Some(found_module_ref) =
+        world_builder.find_module(import_func_id.module.as_symbol())
+    {
+        found_module_ref
+    } else {
+        world_builder
+            .declare_module(import_func_id.module)
+            .expect("failed to create a module for imports")
+    };
+    let mut import_module_builder = ModuleBuilder::new(import_module_ref);
+    let import_func_ref = import_module_builder
+        .define_function(import_func_id.function, import_sig.clone())
+        .expect("failed to create an import function");
+    let results = transform_miden_abi_call(
+        import_func_ref,
+        import_func_id,
+        args.as_slice(),
+        &mut func_builder,
+    );
+
+    let exit_block = func_builder.create_block();
+    func_builder.append_block_params_for_function_returns(exit_block);
+    func_builder.ins().br(exit_block, results, span);
+    func_builder.seal_block(exit_block);
+    func_builder.switch_to_block(exit_block);
+    func_builder.ins().ret(None, span).expect("failed ret");
+
+    CallableFunction {
+        wasm_id: synth_func_id,
+        function_ref: Some(func_ref),
+        signature: synth_func_sig,
+    }
 }
