+++ tests/integration/expected/mem_intrinsics_heap_base.hir
